Only report debug messages in test, on Elixir 1.11+
Logger.put_module_level/2 is not supported on Elixir 1.10.

We only really need these messages in test, so we can completely compile
them out in other situations.

We still need the module level stuff, because somebody might use
PathGlob in their own test environment and not want to see the messages.